+++ guillotina/fields/patch.py
@@ -20,6 +20,7 @@ class PatchField(schema.Field):
     def __init__(self, field, *args, **kwargs):
         self.field = field
         super().__init__(*args, **kwargs)
+        self.required = field.required
 
     def set(self, obj, value):
         bound_field = self.field.bind(obj)